did the filesystem swap it seems to work when it wants to, I don't know if it's a me thing or a program thing, still it works seems to be faster, and smaller too best of both worlds hopefully this precived fragillity dosn't bite me in the ass

+++ Src/Err/err.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void err(int32_t code);
+void err(int16_t code);
 
 #endif
